refactor: restructuring unreadable codes
