+++ lib/components/curved-left-shadow.dart
