+++ lib/components/curved-right-shadow.dart
