+++ lib/components/curved-right.dart
